update questionnaire form

+++ src/app/api/configWeb/route.ts
@@ -1,20 +1,25 @@
-import axios from 'axios';
-import { getToken } from 'next-auth/jwt';
-import { NextRequest, NextResponse } from 'next/server';
+import axios from "axios";
+import { cookies } from "next/headers";
+import { NextRequest, NextResponse } from "next/server";
 
 const secret = process.env.NEXTAUTH_SECRET;
 
 export async function GET(req: NextRequest) {
-  const token = await getToken({ req, secret });
   const u = new URL(req.url);
-  const endpoint = u.searchParams.get('endpoint') as string;
+  const endpoint = u.searchParams.get("endpoint") as string;
   const url = new URL(`${process.env.API_WEB_URL}/${endpoint}`);
 
   try {
+    const acc_token = cookies().get("survey_access_token")?.value;
+
+    if (!acc_token) {
+      throw new Error("Access token not found in cookies");
+    }
+
     const response = await axios.get(url.toString(), {
       headers: {
-        'Content-Type': 'application/json',
-        Authorization: `Bearer ${token?.accessToken}`,
+        "Content-Type": "application/json",
+        Authorization: `Bearer ${acc_token}`,
       },
     });
 
@@ -26,66 +31,99 @@ export async function GET(req: NextRequest) {
 }
 
 export async function POST(req: NextRequest) {
-  const token = await getToken({ req, secret });
-  const { body, endpoint } = (await req.json()) as { body: object; endpoint: string };
+  const { body, endpoint } = (await req.json()) as {
+    body: object;
+    endpoint: string;
+  };
+
+  console.log("endpoint: ", endpoint);
+
+  console.log("body: ", body);
+
   const url = new URL(`${process.env.API_WEB_URL}/${endpoint}`);
 
   try {
+    const acc_token = cookies().get("survey_access_token")?.value;
+
+    if (!acc_token) {
+      throw new Error("Access token not found in cookies");
+    }
+
     const response = await axios.post(
       url.toString(),
       { ...body },
       {
         headers: {
-          'Content-Type': 'application/json',
-          Authorization: `Bearer ${token?.accessToken}`,
+          "Content-Type": "application/json",
+          Authorization: `Bearer ${acc_token}`,
         },
       },
     );
     return NextResponse.json(response.data);
   } catch (error) {
     console.error(`Error fetching ${endpoint}:`, error);
-    return NextResponse.json({ error: `Failed to fetch ${endpoint}` }, { status: 500 });
+    return NextResponse.json(
+      { error: `Failed to fetch ${endpoint}` },
+      { status: 500 },
+    );
   }
 }
 
 export async function PUT(req: NextRequest) {
-  const token = await getToken({ req, secret });
-  const { body, endpoint } = (await req.json()) as { body: object; endpoint: string };
+  const { body, endpoint } = (await req.json()) as {
+    body: object;
+    endpoint: string;
+  };
   const url = new URL(`${process.env.API_WEB_URL}/${endpoint}`);
 
   try {
+    const acc_token = cookies().get("survey_access_token")?.value;
+
+    if (!acc_token) {
+      throw new Error("Access token not found in cookies");
+    }
     const response = await axios.put(
       url.toString(),
       { ...body },
       {
         headers: {
-          'Content-Type': 'application/json',
-          Authorization: `Bearer ${token?.accessToken}`,
+          "Content-Type": "application/json",
+          Authorization: `Bearer ${acc_token}`,
         },
       },
     );
     return NextResponse.json(response.data);
   } catch (error) {
     console.error(`Error updating ${endpoint}:`, error);
-    return NextResponse.json({ error: `Failed to update ${endpoint}` }, { status: 500 });
+    return NextResponse.json(
+      { error: `Failed to update ${endpoint}` },
+      { status: 500 },
+    );
   }
 }
 
 export async function DELETE(req: NextRequest) {
-  const token = await getToken({ req, secret });
   const { endpoint } = (await req.json()) as { endpoint: string };
   const url = new URL(`${process.env.API_WEB_URL}/${endpoint}`);
 
   try {
+    const acc_token = cookies().get("survey_access_token")?.value;
+
+    if (!acc_token) {
+      throw new Error("Access token not found in cookies");
+    }
     const response = await axios.delete(url.toString(), {
       headers: {
-        'Content-Type': 'application/json',
-        Authorization: `Bearer ${token?.accessToken}`,
+        "Content-Type": "application/json",
+        Authorization: `Bearer ${acc_token}`,
       },
     });
     return NextResponse.json(response.data);
   } catch (error) {
     console.error(`Error deleting at ${endpoint}:`, error);
-    return NextResponse.json({ error: `Failed to delete at ${endpoint}` }, { status: 500 });
+    return NextResponse.json(
+      { error: `Failed to delete at ${endpoint}` },
+      { status: 500 },
+    );
   }
 }
